fix: parse_request()'s signature
parse_request() altered to receive an HTTPRequest by reference
(along with an std::string) and return a std::string. The function
also became a member of MsgProcessor. It must be further developed
to parse the incoming request and return the appropriate error res-
ponses, or an empty string if there is no error during parsing.